Create a JSON config file

Create a JSON config file

Usage

config_setup(apikey, companyname, conffile, verbose = TRUE)

Arguments

apikey

API key to access the bambooHR API. User will be prompted to enter one if not supplied

companyname

This is the subdomain used to access BambooHR. If you access BambooHR at https://mycompany.bamboohr.com, then companyname will be "mycompany"

conffile

A string giving the path where the configuration file will be written. This will automatically attempt to load from the default location ~/.bambooHR_user_config.json (which can be changed via a bambooHR.config_file) entry in options).

verbose

whether to print path of conf file used, passed to use_config

Examples

if (FALSE) {
config_setup("secret api key", "cool company")
}